refactor: rename signozauditexporter to signozclickhouseauditexporter
Follow OTel contrib convention: package name includes "exporter" suffix,
component type does not. Type is now "signozclickhouseaudit".
