+++ exporter/signozclickhouseauditexporter/exporter.go
@@ -1,4 +1,4 @@
-package signozauditexporter
+package signozclickhouseauditexporter
 
 import (
 	"context"
